Split test_events into test_events and test_subevents
